refactor(all): Car file path change
