--- src/main/java/racingcar/RacingCar.java
@@ -1,2 +1,0 @@
-package racingcar;public class RacingCar {
-}
