--- src/main/java/racingcar/domain/Car.java
@@ -1,48 +1,0 @@
-package racingcar.domain;
-
-public class Car {
-
-    private static final int MIN_NAME_LENGTH = 1;
-    private static final int MAX_NAME_LENGTH = 5;
-    private static final int MOVE_BOUND = 4;
-
-    private String name;
-    private int position = 0;
-
-    public Car(String name) {
-        validateNameLength(name);
-        this.name = name;
-    }
-
-    private void validateNameLength(String name) {
-        if (name.length() > MAX_NAME_LENGTH || name.length() < MIN_NAME_LENGTH) {
-            throw new IllegalArgumentException("자동차 이름은 1글자 이상 5글자 이하입니다.");
-        }
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getPosition() {
-        return position;
-    }
-
-    public void move(int randomNumber) {
-        if (randomNumber >= MOVE_BOUND) {
-            position++;
-        }
-    }
-
-    public String positionToString() {
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < position; i++) {
-            sb.append("-");
-        }
-        return sb.toString();
-    }
-
-    public boolean isMaxPosition(int maxPosition) {
-        return position == maxPosition;
-    }
-}
